refactor: simplify clean before install directory check

+++ pm
@@ -127,15 +127,13 @@ class Package:
         assert m is not None
         package_dir = os.path.join(m.datadir, self.name)
         os.makedirs(package_dir, exist_ok=True)
-        if self.clean_before_install:
-            if os.listdir(package_dir):
-                if m._interactive_ask("Clean", f"Directory {package_dir} is not empty, remove all files before install?"):
-                    for entry in os.listdir(package_dir):
-                        full_path = os.path.join(package_dir, entry)
-                        if os.path.isdir(full_path):
-                            shutil.rmtree(full_path)
-                        else:
-                            os.remove(full_path)
+        if self.clean_before_install and len(os.listdir(package_dir)) > 0 and m._interactive_ask("Clean before install", package_dir):
+            for entry in os.listdir(package_dir):
+                full_path = os.path.join(package_dir, entry)
+                if os.path.isdir(full_path):
+                    shutil.rmtree(full_path)
+                else:
+                    os.remove(full_path)
         try:
             m.log(f"cd {package_dir}")
             os.chdir(package_dir)
